refactor(utils/logger): added init, error and formatter modules and added tests
